+++ hat/assets/js/apps/Iaso/domains/payments/hooks/requests/useGetPotentialPayments.tsx
@@ -1,9 +1,11 @@
 /* eslint-disable camelcase */
 import { UseQueryResult } from 'react-query';
+import moment from 'moment';
 import { makeUrlWithParams } from '../../../../libs/utils';
 import { getRequest } from '../../../../libs/Api';
 import { useSnackQuery } from '../../../../libs/apiHooks';
 import { PotentialPaymentParams, PotentialPaymentPaginated } from '../../types';
+import { apiDateFormat } from '../../../../utils/dates';
 
 const apiUrl = '/api/potential_payments/';
 
@@ -21,8 +23,16 @@ const getPotentialPayments = (options: PotentialPaymentParams) => {
         order: options.order || 'user__last_name',
         limit: options.pageSize || 20,
         page,
-        change_requests__created_at_after,
-        change_requests__created_at_before,
+        change_requests__created_at_after: change_requests__created_at_after
+            ? moment(change_requests__created_at_after, 'L').format(
+                  apiDateFormat,
+              )
+            : undefined,
+        change_requests__created_at_before: change_requests__created_at_before
+            ? moment(change_requests__created_at_before, 'L').format(
+                  apiDateFormat,
+              )
+            : undefined,
         parent_id,
         forms,
         users,
